% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.validate_input_paths}
\alias{.validate_input_paths}
\title{Check that paths in a config list point to actual rds files}
\usage{
.validate_input_paths(config_list)
}
\arguments{
\item{config_list}{A List. A model configuration list.}
}
\description{
Given a set of input parameters run a housing-led for small areas
}
