% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regross.R
\name{regross}
\alias{regross}
\title{Regrosser}
\usage{
regross(
  base_in,
  base_out,
  target_net,
  col_inflow = "inflow",
  col_outflow = "outflow",
  col_target = "net_target"
)
}
\arguments{
\item{base_in}{Dataframe}

\item{base_out}{Dataframe}

\item{target_net}{Dataframe}

\item{col_inflow}{String}

\item{col_outflow}{String}

\item{col_target}{String}
}
\description{
Regrosser
}
