% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_core.R
\name{trend_core}
\alias{trend_core}
\title{Project from a starting population using a CCM methodology to
produce an output population and components}
\usage{
trend_core(
  start_population,
  fertility_rates,
  mortality_rates,
  out_rates,
  in_flows,
  projection_year,
  constraint_list
)
}
\arguments{
\item{start_population}{A dataframe. The starting population}

\item{fertility_rates}{A dataframe. Fertility rates for the \code{projection_year}}

\item{mortality_rates}{A dataframe. Mortality rates for the \code{projection_year}}

\item{out_rates}{A dataframe. Out migration rates for the \code{projection_year}}

\item{in_flows}{A dataframe. In migration flows for the \code{projection_year}}

\item{projection_year}{Numeric. The year being projected}

\item{constraint_list}{A list of constraints.}
}
\value{
A list of projected components
}
\description{
Project from a starting population using a CCM methodology to
produce an output population and components
}
